src/vwap_levels/pyproject.toml
[project]
name = "vwap-levels"
version = "0.1.0"
description = "`vwap-levels"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "tinkoff-investments>=0.2.0b115",
]
